+++ tests/testthat/test-epw.R
@@ -519,13 +519,10 @@ test_that("download_weather()", {
 
     # download weather
     expect_message({path_epw <- with_verbose(
-        download_weather("USA_CA_San.Francisco.Intl.AP.724940_TMY3", ask = FALSE, type = "epw", dir = tempdir()))}
-    )
+        download_weather("USA_CA_San.Francisco.Intl.AP.724940_TMY3", ask = FALSE, type = "epw", dir = tempdir(), max_match = 2)
+    )})
     expect_message({path_epw <- with_verbose(
-        download_weather("USA_CA_San.Francisco.Intl.AP.724940_TMY3", ask = FALSE, type = "all", dir = tempdir()))}
-    )
-    expect_message({path_epw <- with_verbose(
-        download_weather("USA_CA_San.Francisco.Intl.AP.724940_TMY3", ask = FALSE, type = "ddy", dir = tempdir()))}
-    )
+        download_weather("Nanjing.*TMYx.2004-2018", "Nanjing", ask = FALSE, type = "epw", dir = tempdir(), max_match = 1)
+    )})
 })
 # }}}