+++ NEWS.md
@@ -3,6 +3,14 @@
 ## New features
 
 * `Idd$path()` is added to show the path of IDD parsed (#442).
+* Add epw data sources from [climate.onebuilding.org](http://climate.onebuilding.org/)
+  for `download_weather()`. `type` will always be `"all"` for those sources (#453).
+* Add a new `"stat"` option in `type` in `download_weather()` (#453).
+
+## Minor changes
+
+* When `type` is `"all"` in `download_weather()`, the ZIP file will be
+  downloaded instead of downloading both `EPW` and `DDY` files (#453).
 
 ## Bug fixes